refactor: simplify secure vault keygen peer discovery step
